Bug #651187 remove ALC/SFT AF, lib_ctrl.c build warning
[root cause  ] remove waring
[changes     ] warning
[side effects] NA
[self test   ] 9850ka_1h10, no warning in ALC AF
[reviewers   ] wuyi.shen
[change_type ] feature_bugfix
[tag_product ] common

+++ isp2.1/af/alc_af/ALC_AF_Interface.c
@@ -38,7 +38,7 @@
 #include <utils/Log.h> //for test
 /* ------------------------------------------------------------------------ */
 
-void Al_AF_GetFv(alc_af_handle_t handle,unsigned long dwValue[150])
+void Al_AF_GetFv(alc_af_handle_t handle, uint32_t dwValue[150])
 {
 	struct alc_af_context *af_cxt = (struct alc_af_context *)handle;
